Generated the function for the computer to randomly generate an option in the console.
There are no parameter to the function itself, once function is called, random number is generated and the function carrys out a if else where the conditions are dependent on the random number generated, since there are three options each option was given a 1/3 section of 0 to 1 assigned to it to make the function return a value of the form of a string from the choices: rock, paper, scissors. Next step would be assiging a variable (computerChoice) that is equal to the return value of this function.