Made the if else statement a function and wrapped that in its own function
wrapped the playRound function in a startRound function that has the parameters of playRound for the computer's and user's inputs in the if else statement auto set to the prompt function asking user to type a selection, and the computerChoice variable. This causes the prompt to only appear after the user types in the function in the console and thus each round in the future series can have different inputs. Thus, I will need to sub the computerChoice variable for the getComputerChoice function in this startRound function as a parameter entry for the playRound function that way a new value is generated each round in a series that way the computer choice is not the same every loop.