Added a new function that keeps tally of the user and computer score
Based off of reading the roundResult variable in the playRound function to then either increment a integer value for either a variable called userScore or computerScore and a message that plays via console.log() alerting the user what the current score is. Need to figure out how to output the individual round result first after a round is played then the message with the score update.