changed the condition for the startSeries function and put a reset score feature in the same function
I have set userScore and computerScore = 0 at the top of the startSeries() that way every iteration of the function the scores get reset to 0. I have also changed the condition for the user to see the victory message to just userScore > computerScore until I can figure out how to loop the startRound function in the startSeries function until one of the score == 0 as of right now I just have it set to calling the function nine times and then going through the if else statement.