Changed the startSeries function name to logSeries to be more accurate of current functionality which has also been reimplemented.
Replaced the startRound() in the eventListener to the logSeries() to keep track of the series score until either the computer or the user reach a score of 5 then the logSeries() resets the scores and outputs a victory or defeat message. If neither condition is met it outputs a message to the user to select an option again to continue. Initaially tried to use a while loop in the startSeries() but realized it would take the first input and run the function with that same input until one of the score conditions was met so needed a way to keep track of score, reset scores, all while only updating after the user selects one of the inputs for each round, thus the name change since it no longer starts a series per say but continually updates and resets series without a loop.